create reserved group for all hosts
- temporarily create a reserved group for all hosts. this is needed when you
want to run an ansible command against a host that might not yet be in a group.
This is used for host check and setup. Oracle-Bug:21824287
- handle case when empty inventory has whitespace in it. useful when manually
clearing the inventory file
- add rule preventing adding a group name equal to an existing service name.
Oracle-bug:21831776
- add exception handling to run_cmd to better output cmd failures when an
exception occurs.
- add restriction to avoid pexpect 3.3 in requirements. That version has a
bug where it will fail sudo commands.
- change unittest kolla_cli etc location to be the user's home location. This
avoids errors when the kolla_cli location is in /etc/kolla
- change format of unittest test_hosts file to remove ip address. no longer
needed.
- remove ndbcluster from group test