update host check to remove obsolete usage of json generator.
Oracle-Bug: 22159340
host check needs to use the new on the fly json generator. Because
the on the fly generator requires the inventory object, the host check is
moved from the host object to the inventory. And because host setup
calls host check, it is also moved to from host to inventory.