create unique deploy id for deployments
As part of using ansible callbacks, this creates a uniquely
named inventory file that will allow the
the ansible callback code to communicate status about a deployment
back to the client api. The directory name will contain
this unique ID.

This is a WIP, this is only changing the inventory code to use
temp directories containing UUIDs.

Jira-Issue: OSTACKDEV-20