Update host destroy and precheck api to take list of hosts
Having 'all' as a hostname is awkward in the api.
This changes the api so that a list of hostnames is
required. 'all' is still valid in the CLI, so no change
to the CLI behavior.

This also fixes bug 22918000 - a hostname
validity check is now done on the host destroy.

Jira-Issue: OPENSTACK-746