fix host & service associations in group object
When hosts or services are added/removed from the
api group object, the group object does not reflect
those changes.

Add doc notes to service and host objects that if group
associations are changed, the object muct be refetched
to reflect the changed associations.

Jira-Issue: OPENSTACK-931
Jira-Issue: OPENSTACK-932
Jira-Issue: OPENSTACK-933