use inventory_samples as name of all-in-one dir
Last Sep-2015, the kolla/ansible/'inventory' directory name
was changed in the kolla rpm spec to 'inventory_samples'. This
was done to avoid confusion with the actual inventory file. However,
the code in the v3.0 cli assumed that the all-in-one file was
still in the inventory directory. With this change the cli will
now look for the all-in-one file in the inventory_samples
directory.

Change-Id: I899104e4dee487458e01644b9f4ef2f8c5d97cfc
Jira-Issue: OPENSTACK-1065