add new cli/api command to init passwords
Passwords in mitaka kolla are now empty. You have to run a kolla-genpwd
command to init the values. But you have to run that as either root or
kolla, which is not desirable.

This adds a new password command to the cli/api to run that script
for the user so they do not have to change to root or kolla.

Bug 24704089
Jira-Issue: None

Change-Id: I16f60635c94f1f44d0489d20362384ee02caf8b5