Added support for upgrade and deploy per service
This adds a --services flag to the kollacli upgrade command.
A list of services can be provided, delimited by commas.

This also adds support for deploy to an individual or set
of services through the api.  Since it may not always work
correctly due to ansible playbooks it is meant to be used
by advanced users that make sure it behaves the way they
want it to.

Also fixing up / commenting out some tests that don't
function at this point.  They will be re-written once
the destroy changes are finished

Also moved verification of host specific deploy only
to compute nodes into the CLI so that advanced use of
the host deploy can be done through the API to
non-compute nodes.

Change-Id: Ia214627e45664d0958a73cb8c940efacd690d5ea
Jira-Issue:OSTACKDEV-148