Updated destroy to use upstream destroy playbook
Due to the change to use the upstream playbook the cli destroy
behavior has changed.  If the --includedata flag is not included
the user will be prompted to enter y/n to indicate it is ok to
delete all the containers and associated data on the host being
destroyed (or all if the host is all).  The stop or kill option
is also no longer relevant.  A new command will be introduced
to support simply stopping containers.

Change-Id: I7076fe8f99a846e43db114dfb232413100dc5049
Jira-Issue: OSTACKDEV-142