Added support for ignoring certain ansible errors.
This change will regex match a set of strings against whatever
errors are encountered.  It will also show a count of the errors
ignored per host and subtract the number of ignored errors from the
failure count.  If the total number of ignored errors equals the
number of failures the job will be considered a success.  This
feature is just meant to be used internally in the ansible
execution code.  It will not be exposed via the API or CLI.

Change-Id: Ib6d3d8e5d3441707346c3ec8f23e6991f71e9dab
Jira-Issue: None