Pass through unlimited -v args to ansible playbook
The existing code would limit the -v args to three and would also
pass through three even if only one or two were passed in on the
command line.  This behavior should be a little more obvious
to the users and manual and cli usage of ansible will get
the same results.

Change-Id: I859178da170941887d3d1b5ecccf966ca912a3ac
Jira-Issue: None