Hard bind cliff dependency and minor unicode fixes
Newer versions of cliff have an issue where unicode strings
are not handled correctly in output so we hard bind the
requirements to a version of cliff we know doesn't have
the problem.  Also we were only doing the unicode translation
when outputing to a table where there were are few other
output formats where the translation was also appropriate.

Change-Id: I04f7988f5ff01f7434687fade5465b87199b86ba