Re-worked the way service association is handled
Prior to this change it was impossible to have more
than a single level of service nesting and
sub-services that were directly associated to groups
would not have been represented properly in the inventory
file we generate during the ansible playbook run.
Now we can next services to whatever degree necessary
and groups can be associated to any service regardless
of nesting depth.

Change-Id: I3f1635585adbceb47cca3220348332ab522b49c4
Jira-Issue: OPENSTACK-1613