Support modification of non string properties
Support modification of non string properties.  Introduced
the concept of types to the AnsibleProperty object and when
a property is set if the type is not str or None then
the value will be loaded via yaml and processed into the
appropriate python type.  At write time it is then written
out via yaml dump rather than always writing it out as a
string.

Change-Id: Iab1d10430cb919cf968e4357c93e23ed360bf338
Jira-Issue: OPENSTACK-1698