Run certificate init playbook as admin user
Previously, the generated certificate files were owned by root which
prevented it from being easily removed later.  This change adds the
capability of setting the --become-user flag to ansible-playbook which
allows the certificate generation playbook to run as the admin user so
that the resulting certificate files are owned by the admin user.

Change-Id: I3c5cf1894b88474f6fd89a27c52bd2d1257e3ebe