Add unit tests for host commands
As part of getting away from using calls to the shell in the
tests, we will need code coverage of the commands library which
will no longer be used when the shell is no longer called. This
change is the first of serveral to provide unnit test coverage
for commands/host.

As a result of writing these tests, I needed to tweak a few things,
some to make mocking possible, and fix some minor issues:

- move delete confirmation prompt to its own method so it can
  be mocked.
- remove extra space from failed Host check message
- fix return code to 0 in Host setup, when the host is already setup.
- in shell, move inventory check to a method so it can be mocked.

Change-Id: I63cb1afd5313959a6fdda11e9c2b03317c60197a