Add setup script for kolla-cli / kolla-ansible
Created new cli-setup.py script that automates the previously
manual process to create the directories needed for the kolla-cli
and kolla-ansible.  It also improves on the process by leaving the
globals.yml file in place in /etc/kolla and just creating a link
to the group_vars/__GLOBALS__ so that both the kolla-ansible and
kolla-cli commands will work correctly, rather than having their
functioning being mutually exclusive.  Due to kolla-ansible still
expecting to be run as root for most actions the cli_setup.py
script should also be run via sudo.

Change-Id: I5b59317b2c22d10dd82623a548b64f387b6bc9d3