Add ability to load inventory from a file
It would be helpful to have the ability to populate the
kolla-cli inventory file from any inventory file, not just the
all-in-one file from kolla.

To do this, The config reset command has been updated to take
an optional --inventory path argument that will cause the new
inventory to be generated from that anible inventory file.

While doing this found a few things to fix:

- renamed allineone.py & AllInOne to AnsibleInventory to better reflect
what they are now doing.

- In kolla_actions, preserve the empty globals.yml file (otherwise
the soft linkage is broken) and fix the kolla-cli directory name.

Change-Id: Ib250c5de8e00818b41c1a4539aba8c165a4ad2e2