Add import inventory command & api
This undoes a part of a prior checkin
(Ib250c5de8e00818b41c1a4539aba8c165a4ad2e2) that added an inventory
import option to the "config reset" command. That turned out to
be too limiting, so this is changing config reset back to the way it
was and adds a new import inventory command. This command
could be augmented in the future to add the ability to import
group variables, host variables or passwords from files.

Change-Id: If930ab19f9c666d9d025c63cf10c1fad758a7247
Related-to: Ib250c5de8e00818b41c1a4539aba8c165a4ad2e2