Fixing test failures
It seems that in newer ansible versions the error strings have
changed and we are no longer able to use the hard coded
"UNREACHABLE" string in our stop test.  For now changing the
test to just skip the string investigation.  Also updating some
package dependencies to match those in kolla-ansible.

We need the 3.5 to 3.6 migration along with the UNREACHABLE
string change together to clean up the test failures  Neither
can be fixed without the other.  Thanks zhulingjie!!

Co-Authored-By: zhulingjie <easyzlj@gmail.com>
Change-Id: I347cda4bbdc6ed7dc4857cb1865da578670235fb